pyproject.toml
[project]
name = "digitise-lecture-notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.3",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "openai>=1.58.1",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.4",
]
